requirements.txt
boto3
botocore
PyYAML>=6.0
